requirements.txt
quart
hypercorn
google-genai
groq
pinecone-client
python-dotenv
requests
